pyproject.toml
[tool.poetry]
name = "headlight"
description = "A database migration toolkit."
version = "0.1.8"
authors = ["Alex Oleshkevich <alex.oleshkevich@gmail.com>"]
license = "MIT"
readme = "README.md"
homepage = "https://github.com/alex-oleshkevich/headlight"
repository = "https://github.com/alex-oleshkevich/headlight"
documentation = "https://github.com/alex-oleshkevich/headlight"
keywords = []
classifiers = [
    "Development Status :: 5 - Production/Stable",
    "Environment :: Web Environment",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Typing :: Typed",
]

[tool.poetry.dependencies]
python = "^3.10.0"
click = "^8.1.3"
tomlkit = "^0.11.1"
psycopg2-binary = { version = "^2.9.3", optional = true }

[tool.poetry.dev-dependencies]
pytest = "^7.1.0"
pytest-asyncio = "^0.18.0"
pytest-cov = "^3.0"
black = "^22.6.0"
mypy = "^0.971"
flake8 = "^4.0.1"

[tool.poetry.extras]
postgresql = ["psycopg2-binary"]

[tool.poetry.scripts]
headlight = 'headlight.console:main'

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.black]
target-version = ["py310"]
line_length = 120
extend-exclude = true
exclude = '''
(
    asv_bench/env
  | \.egg
  | \.git
  | \.hg
  | \.mypy_cache
  | \.nox
  | \.tox
  | \.venv
  | _build
  | buck-out
  | build
  | dist
  | setup.py
)
'''


[tool.isort]
combine_as_imports = true
include_trailing_comma = true
known_standard_library = "dataclasses,typing_extensions"
known_first_party = "headlight"
known_third_party = ["pkg_resources", "toml"]
known_future_library = "__future__"
line_length = 120
multi_line_output = 3
profile = "black"
use_parentheses = true

[tool.coverage.run]
branch = true
source = ["headlight"]
omit = ["tests/*", ".venv/*", ".git/*", "*/__main__.py"]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "raise NotImplementedError"
]

[tool.mypy]
disallow_untyped_defs = true
ignore_missing_imports = true
#pretty=true
show_error_codes = true
files = ["headlight", "tests"]

[tool.pytest.ini_options]
minversion = "7.1"
python_files = ["tests.py", "test_*.py", "*_tests.py"]
norecursedirs = [
    "node_modules", "frontend", "storage", "dist", ".git",
    "*/migrations/*", "*/static/*", "docs", ".venv"
]
addopts = "--tb=short -s --no-cov-on-fail"
testpaths = ["tests"]
filterwarnings = [
    "ignore::UserWarning",
    "ignore::DeprecationWarning",
]
